refactor Basic/BasicWL
