+++ src/DataType/IncludedSupplyChainTradeLineItem.php
@@ -1,8 +1,7 @@
 <?php
 
-namespace Tiime\CrossIndustryInvoice\Basic;
+namespace Tiime\CrossIndustryInvoice\DataType;
 
-use Tiime\CrossIndustryInvoice\DataType\AssociatedDocumentLineDocument;
 use Tiime\CrossIndustryInvoice\DataType\Basic\SpecifiedLineTradeDelivery;
 use Tiime\CrossIndustryInvoice\DataType\Basic\SpecifiedTradeProduct;
