+++ tests/CIIBasicTest.php
@@ -4,10 +4,6 @@ namespace Tiime\CrossIndustryInvoice\Tests;
 
 use PHPUnit\Framework\TestCase;
 use Tiime\CrossIndustryInvoice\Basic\CrossIndustryInvoice;
-use Tiime\CrossIndustryInvoice\Basic\IncludedSupplyChainTradeLineItem;
-use Tiime\CrossIndustryInvoice\Basic\SpecifiedLineTradeAgreement;
-use Tiime\CrossIndustryInvoice\Basic\SpecifiedLineTradeSettlement;
-use Tiime\CrossIndustryInvoice\Basic\SupplyChainTradeTransaction;
 use Tiime\CrossIndustryInvoice\DataType\ActualDeliverySupplyChainEvent;
 use Tiime\CrossIndustryInvoice\DataType\AssociatedDocumentLineDocument;
 use Tiime\CrossIndustryInvoice\DataType\Basic\ApplicableTradeTax;
@@ -15,6 +11,7 @@ use Tiime\CrossIndustryInvoice\DataType\Basic\BilledQuantity;
 use Tiime\CrossIndustryInvoice\DataType\Basic\SpecifiedLineTradeDelivery;
 use Tiime\CrossIndustryInvoice\DataType\Basic\SpecifiedTradeProduct;
 use Tiime\CrossIndustryInvoice\DataType\Basic\SpecifiedTradeSettlementLineMonetarySummation;
+use Tiime\CrossIndustryInvoice\DataType\Basic\SupplyChainTradeTransaction;
 use Tiime\CrossIndustryInvoice\DataType\BasicWL\ApplicableHeaderTradeAgreement;
 use Tiime\CrossIndustryInvoice\DataType\BasicWL\ApplicableHeaderTradeDelivery;
 use Tiime\CrossIndustryInvoice\DataType\BasicWL\ApplicableHeaderTradeSettlement;
@@ -30,12 +27,15 @@ use Tiime\CrossIndustryInvoice\DataType\DespatchAdviceReferencedDocument;
 use Tiime\CrossIndustryInvoice\DataType\DocumentIncludedNote;
 use Tiime\CrossIndustryInvoice\DataType\ExchangedDocumentContext;
 use Tiime\CrossIndustryInvoice\DataType\GuidelineSpecifiedDocumentContextParameter;
+use Tiime\CrossIndustryInvoice\DataType\IncludedSupplyChainTradeLineItem;
 use Tiime\CrossIndustryInvoice\DataType\IssueDateTime;
 use Tiime\CrossIndustryInvoice\DataType\LocationGlobalIdentifier;
 use Tiime\CrossIndustryInvoice\DataType\NetPriceProductTradePrice;
 use Tiime\CrossIndustryInvoice\DataType\OccurrenceDateTime;
 use Tiime\CrossIndustryInvoice\DataType\SellerGlobalIdentifier;
 use Tiime\CrossIndustryInvoice\DataType\ShipToTradeParty;
+use Tiime\CrossIndustryInvoice\DataType\SpecifiedLineTradeAgreement;
+use Tiime\CrossIndustryInvoice\DataType\SpecifiedLineTradeSettlement;
 use Tiime\CrossIndustryInvoice\DataType\SpecifiedTaxRegistration;
 use Tiime\CrossIndustryInvoice\DataType\URIUniversalCommunication;
 use Tiime\EN16931\DataType\CountryAlpha2Code;
